Get settings files compiling
